requirements.txt
kafka-python
requests
pandas
pyarrow